requirements.txt
torch
opencv-python
Pillow
pandas
tqdm
transformers
scikit-learn
timm
numpy
einops